+++ cs104/config.go
@@ -45,7 +45,7 @@ type Config struct {
 	// I-frames 发送未收到确认的帧数上限， 一旦达到这个数，将停止传输
 	// "k" 范围[1, 32767] 默认 12.
 	// See IEC 60870-5-104, subclause 5.5.
-	SendUnackLimitK uint
+	SendUnackLimitK uint16
 
 	// 帧接收确认最长超时时间，超过此时间立即关闭连接。
 	// "t₁" 范围[1, 255]s 默认 15s.
@@ -55,7 +55,7 @@ type Config struct {
 	// 接收端最迟在接收了w次I-frames应用规约数据单元以后发出认可。 w不超过2/3k(2/3 SendUnackMax)
 	// "w" 范围 [1， 32767] 默认 8.
 	// See IEC 60870-5-104, subclause 5.5.
-	RecvUnackLimitW uint
+	RecvUnackLimitW uint16
 
 	// 发送一个接收确认的最大时间，实际上这个框架1秒内发送回复
 	// "t₂" 范围[1, 255]s 默认 10s
